Move emails not associated with individual users to ReportMailer
