+++ config/locales/report_mailer/en.yml
@@ -1,0 +1,7 @@
+---
+en:
+  report_mailer:
+    sps_over_quota_limit:
+      help: ''
+      info: There are SPs that have gone over their quota limit.
+      subject: Some SPs have reached their quota limit
