+++ config/locales/user_mailer/en.yml
@@ -178,10 +178,6 @@ en:
         you can ignore this message.
       link_text: Go to %{app}
       reset_password: If you can't remember your password, go to %{app} to reset it.
-    sps_over_quota_limit:
-      help: ''
-      info: There are SPs that have gone over their quota limit.
-      subject: Some SPs have reached their quota limit
     undeliverable_address:
       call_to_action: Please sign into %{link} and follow instructions.
       help: ''
