+++ spec/mailers/user_mailer_spec.rb
@@ -453,25 +453,6 @@ describe UserMailer, type: :mailer do
     end
   end
 
-  describe 'sps_over_quota_limit' do
-    let(:mail) { UserMailer.sps_over_quota_limit(email_address.email) }
-
-    it_behaves_like 'a system email'
-
-    it 'sends to the current email' do
-      expect(mail.to).to eq [email_address.email]
-    end
-
-    it 'renders the subject' do
-      expect(mail.subject).to eq t('user_mailer.sps_over_quota_limit.subject')
-    end
-
-    it 'renders the body' do
-      expect(mail.html_part.body).
-        to have_content(strip_tags(t('user_mailer.sps_over_quota_limit.info')))
-    end
-  end
-
   def strip_tags(str)
     ActionController::Base.helpers.strip_tags(str)
   end
